requirements.txt
fastapi
uvicorn
pydantic
sqlglot==27.6.0
python-dotenv
pytest
httpx
psycopg2-binary
requests
prometheus-client
slowapi